LPS-145450 Rename Google Cloud translator packages
