+++ modules/apps/translation/translation-google-cloud-translator/bnd.bnd
@@ -1,6 +1,6 @@
-Bundle-Name: Liferay Translation Google Cloud Translator
-Bundle-SymbolicName: com.liferay.translation.google.cloud.translator
-Bundle-Version: 1.0.11
+Bundle-Name: Liferay Translation Translator Google Cloud
+Bundle-SymbolicName: com.liferay.translation.translator.google.cloud
+Bundle-Version: 1.0.0
 Import-Package:\
 	!com.google.appengine.api.*,\
 	!com.google.protobuf.*,\
